pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paic"
version = "0.1.0"
description = "Prisma Access IP Console — multi-tenant egress IP management"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "apscheduler>=3.10.4",
    "netaddr>=1.2.1",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.3",
    "prometheus-client>=0.25.0",
    "aiosmtplib>=5.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-netaddr>=1.2.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/paic"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[project.scripts]
paic = "paic.cli:main"

[tool.coverage.run]
source = ["src/paic"]
omit = [
    "src/paic/cli.py",
]

[tool.coverage.report]
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "aiosmtpd>=1.4.6",
]
